requirements.txt
torch
sentence-transformers
scikit-learn
faiss-cpu
transformers
streamlit